lint + single plugins to avoid duplication
