+++ projects/safe/src/lib/components/ui/bar-chart/bar-chart.component.ts
@@ -7,10 +7,11 @@ import {
   ChartType,
 } from 'chart.js';
 import { BaseChartDirective } from 'ng2-charts';
-import drawUnderlinePlugin from '../../../utils/graphs/plugins/underline';
+import drawUnderlinePlugin from '../../../utils/graphs/plugins/underline.plugin';
 import DataLabelsPlugin from 'chartjs-plugin-datalabels';
 import { parseFontOptions } from '../../../utils/graphs/parseFontString';
 import { addTransparency } from '../../../utils/graphs/addTransparency';
+import whiteBackgroundPlugin from '../../../utils/graphs/plugins/background.plugin';
 
 /**
  * Interface of chart title.
@@ -32,21 +33,6 @@ interface ChartLegend {
 }
 
 /**
- * Custom plugin for having a white background on the charts
- */
-const whiteBackgroundPlugin = {
-  id: 'customCanvasBackgroundColor',
-  beforeDraw: (chart: any) => {
-    const {ctx} = chart;
-    ctx.save();
-    ctx.globalCompositeOperation = 'destination-over';
-    ctx.fillStyle = 'white';
-    ctx.fillRect(0, 0, chart.width, chart.height);
-    ctx.restore();
-  }
-};
-
-/**
  * Bar/Column chart component, based on chart.js component.
  */
 @Component({
@@ -58,7 +44,7 @@ export class SafeBarChartComponent implements OnChanges {
   public plugins: ChartComponentLike[] = [
     drawUnderlinePlugin,
     DataLabelsPlugin,
-    whiteBackgroundPlugin
+    whiteBackgroundPlugin,
   ];
   private usePercentage = false;
   private showValueLabels: false | 'percentage' | 'value' = false;
