+++ projects/safe/src/lib/components/ui/line-chart/line-chart.component.ts
@@ -9,9 +9,10 @@ import {
 } from 'chart.js';
 import { BaseChartDirective } from 'ng2-charts';
 import DataLabelsPlugin from 'chartjs-plugin-datalabels';
-import drawUnderlinePlugin from '../../../utils/graphs/plugins/underline';
+import drawUnderlinePlugin from '../../../utils/graphs/plugins/underline.plugin';
 import { parseFontOptions } from '../../../utils/graphs/parseFontString';
 import { addTransparency } from '../../../utils/graphs/addTransparency';
+import whiteBackgroundPlugin from '../../../utils/graphs/plugins/background.plugin';
 
 /**
  * Interface containing the settings of the chart title
@@ -33,21 +34,6 @@ interface ChartLegend {
 }
 
 /**
- * Custom plugin for having a white background on the charts
- */
-const whiteBackgroundPlugin = {
-  id: 'customCanvasBackgroundColor',
-  beforeDraw: (chart: any) => {
-    const {ctx} = chart;
-    ctx.save();
-    ctx.globalCompositeOperation = 'destination-over';
-    ctx.fillStyle = 'white';
-    ctx.fillRect(0, 0, chart.width, chart.height);
-    ctx.restore();
-  }
-};
-
-/**
  * Uses chart.js to render the data as a line chart
  */
 @Component({
@@ -59,7 +45,7 @@ export class SafeLineChartComponent implements OnChanges {
   public plugins: ChartComponentLike[] = [
     drawUnderlinePlugin,
     DataLabelsPlugin,
-    whiteBackgroundPlugin
+    whiteBackgroundPlugin,
   ];
   private showValueLabels = false;
   private min = Infinity;
@@ -87,9 +73,9 @@ export class SafeLineChartComponent implements OnChanges {
     },
     elements: {
       line: {
-        spanGaps: true
-      }
-    }
+        spanGaps: true,
+      },
+    },
   };
 
   public chartType: ChartType = 'line';
@@ -171,7 +157,7 @@ export class SafeLineChartComponent implements OnChanges {
           position: get(this.title, 'position', 'top'),
           color: titleColor,
           font: fontOptions,
-        }
+        },
       },
     } as ChartOptions;
