+++ projects/safe/src/lib/components/ui/pie-donut-chart/pie-donut-chart.component.ts
@@ -3,9 +3,10 @@ import { ChartComponentLike, ChartConfiguration, ChartData } from 'chart.js';
 import { BaseChartDirective } from 'ng2-charts';
 import { get, flatten } from 'lodash';
 import { parseFontOptions } from '../../../utils/graphs/parseFontString';
-import drawUnderlinePlugin from '../../../utils/graphs/plugins/underline';
+import drawUnderlinePlugin from '../../../utils/graphs/plugins/underline.plugin';
 import DataLabelsPlugin from 'chartjs-plugin-datalabels';
 import { addTransparency } from '../../../utils/graphs/addTransparency';
+import whiteBackgroundPlugin from '../../../utils/graphs/plugins/background.plugin';
 
 /**
  * Interface containing the settings of the chart title
@@ -27,21 +28,6 @@ interface ChartLegend {
 }
 
 /**
- * Custom plugin for having a white background on the charts
- */
-const whiteBackgroundPlugin = {
-  id: 'customCanvasBackgroundColor',
-  beforeDraw: (chart: any) => {
-    const {ctx} = chart;
-    ctx.save();
-    ctx.globalCompositeOperation = 'destination-over';
-    ctx.fillStyle = 'white';
-    ctx.fillRect(0, 0, chart.width, chart.height);
-    ctx.restore();
-  }
-};
-
-/**
  * Uses chart.js to render the data as a pie chart
  */
 @Component({
@@ -59,7 +45,7 @@ export class SafePieDonutChartComponent implements OnChanges {
   public plugins: ChartComponentLike[] = [
     drawUnderlinePlugin,
     DataLabelsPlugin,
-    whiteBackgroundPlugin
+    whiteBackgroundPlugin,
   ];
   @Input() chartType: 'pie' | 'doughnut' = 'doughnut';
 
@@ -88,6 +74,7 @@ export class SafePieDonutChartComponent implements OnChanges {
   };
 
   ngOnChanges(): void {
+    console.log(this.options.palette);
     this.showCategoryLabel = get(this.options, 'labels.showCategory', false);
     if (get(this.options, 'labels.showValue', false))
       this.showValueLabels = get(this.options, 'labels.valueType', false);
@@ -96,7 +83,7 @@ export class SafePieDonutChartComponent implements OnChanges {
         (acc: number, curr: any) => acc + curr.field,
         0
       ) || 0;
-    
+
     this.chartData.datasets = this.series.map((x) => ({
       ...x,
       backgroundColor: this.options.palette || undefined,
