+++ projects/safe/src/lib/utils/graphs/plugins/background.plugin.ts
@@ -1,0 +1,16 @@
+/**
+ * Custom plugin for having a white background on the charts
+ */
+const whiteBackgroundPlugin = {
+  id: 'customCanvasBackgroundColor',
+  beforeDraw: (chart: any) => {
+    const { ctx } = chart;
+    ctx.save();
+    ctx.globalCompositeOperation = 'destination-over';
+    ctx.fillStyle = 'white';
+    ctx.fillRect(0, 0, chart.width, chart.height);
+    ctx.restore();
+  },
+};
+
+export default whiteBackgroundPlugin;
